Dedup shared Jira helpers into shared/jira_common.py + Golden Architecture
- Add shared/jira_common.py with eprint, get_jira_instance, CUSTOM_FIELDS
- Add shared/test_jira_common.py (5 tests covering all shared helpers)
- Migrate all 7 Jira scripts to import from shared/ via sys.path.insert
- Remove per-action copies of eprint/get_jira_instance/CUSTOM_FIELDS
- Remove per-action auth tests now covered by shared/test_jira_common.py
- Standardize JIRA_URL ternary to one canonical expression in all action.yml
- Update CLAUDE.md: point CUSTOM_FIELDS at shared/, add Golden Architecture section
  documenting load-bearing constraints (vault blocks, injection guards, fan-outs)
  that must stay per-action, plus conventions for new actions

All 98 tests pass.
