+++ create-integration-ticket/test_create_integration_ticket.py
@@ -15,7 +15,7 @@ from io import StringIO
 sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
 
 from create_integration_ticket import (
-    get_jira_instance, validate_release_ticket, create_integration_ticket,
+    validate_release_ticket, create_integration_ticket,
     link_tickets, main
 )
 from jira.exceptions import JIRAError
@@ -29,65 +29,6 @@ class TestCreateIntegrationTicket(unittest.TestCase):
             'JIRA_USER': 'test_user',
             'JIRA_TOKEN': 'test_token'
         }
-
-    @patch.dict(os.environ, {})
-    def test_get_jira_instance_missing_credentials(self):
-        """Test that get_jira_instance exits when credentials are missing."""
-        with self.assertRaises(SystemExit) as cm:
-            get_jira_instance('https://sonarsource.atlassian.net/')
-        self.assertEqual(cm.exception.code, 1)
-
-    @patch.dict(os.environ, {'JIRA_USER': 'test', 'JIRA_TOKEN': 'token'})
-    @patch('create_integration_ticket.JIRA')
-    def test_get_jira_instance_success_prod(self, mock_jira_class):
-        """Test successful JIRA instance creation for production."""
-        mock_jira = Mock()
-        mock_jira_class.return_value = mock_jira
-
-        jira_client = get_jira_instance('https://sonarsource.atlassian.net/')
-
-        self.assertEqual(jira_client, mock_jira)
-        mock_jira_class.assert_called_once_with(
-            'https://sonarsource.atlassian.net/',
-            basic_auth=('test', 'token'),
-            get_server_info=True
-        )
-
-    @patch.dict(os.environ, {'JIRA_USER': 'test', 'JIRA_TOKEN': 'token'})
-    @patch('create_integration_ticket.JIRA')
-    def test_get_jira_instance_success_sandbox(self, mock_jira_class):
-        """Test successful JIRA instance creation for sandbox."""
-        mock_jira = Mock()
-        mock_jira_class.return_value = mock_jira
-
-        jira_client = get_jira_instance('https://sonarsource-sandbox-608.atlassian.net/')
-
-        self.assertEqual(jira_client, mock_jira)
-        mock_jira_class.assert_called_once_with(
-            'https://sonarsource-sandbox-608.atlassian.net/',
-            basic_auth=('test', 'token'),
-            get_server_info=True
-        )
-
-    @patch.dict(os.environ, {'JIRA_USER': 'test', 'JIRA_TOKEN': 'token'})
-    @patch('create_integration_ticket.JIRA')
-    def test_get_jira_instance_auth_failure(self, mock_jira_class):
-        """Test JIRA instance creation with authentication failure."""
-        mock_jira_class.side_effect = JIRAError(status_code=401, text="Unauthorized")
-
-        with self.assertRaises(SystemExit) as cm:
-            get_jira_instance('https://sonarsource.atlassian.net/')
-        self.assertEqual(cm.exception.code, 1)
-
-    @patch.dict(os.environ, {'JIRA_USER': 'test', 'JIRA_TOKEN': 'token'})
-    @patch('create_integration_ticket.JIRA')
-    def test_get_jira_instance_unexpected_error(self, mock_jira_class):
-        """Test JIRA instance creation with unexpected error."""
-        mock_jira_class.side_effect = Exception("Unexpected error")
-
-        with self.assertRaises(SystemExit) as cm:
-            get_jira_instance('https://sonarsource.atlassian.net/')
-        self.assertEqual(cm.exception.code, 1)
 
     def test_validate_release_ticket_success(self):
         """Test successful release ticket validation."""
