+++ get-jira-release-notes/get_jira_release_notes.py
@@ -10,39 +10,10 @@ import argparse
 import os
 import sys
 from collections import defaultdict
-from jira import JIRA
+
+sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), '..', 'shared'))
+from jira_common import eprint, get_jira_instance
 from jira.exceptions import JIRAError
-
-
-# noinspection DuplicatedCode
-def eprint(*args, **kwargs):
-    """Prints messages to stderr to avoid polluting stdout, which is used for the script's output."""
-    print(*args, file=sys.stderr, **kwargs)
-
-
-def get_jira_instance(jira_url):
-    """Initializes and returns a JIRA client instance."""
-    jira_user = os.environ.get('JIRA_USER')
-    jira_token = os.environ.get('JIRA_TOKEN')
-
-    if not jira_user or not jira_token:
-        eprint("Error: JIRA_USER and JIRA_TOKEN environment variables must be set.")
-        sys.exit(1)
-
-    eprint(f"Connecting to JIRA server at: {jira_url}")
-
-    try:
-        jira_client = JIRA(jira_url, basic_auth=(jira_user, jira_token))
-        jira_client.server_info()
-        eprint("JIRA authentication successful.")
-        return jira_client
-    except JIRAError as e:
-        eprint(f"Error: JIRA authentication failed. Status: {e.status_code}")
-        eprint(f"Response text: {e.text}")
-        sys.exit(1)
-    except Exception as e:
-        eprint(f"An unexpected error occurred during JIRA connection: {e}")
-        sys.exit(1)
 
 
 def get_project_name(jira_client, project_key):
