+++ resolve-ktlo-epic/action.yml
@@ -45,11 +45,9 @@ runs:
       env:
         JIRA_USER: ${{ fromJSON(steps.secrets.outputs.vault).JIRA_USER }}
         JIRA_TOKEN: ${{ fromJSON(steps.secrets.outputs.vault).JIRA_TOKEN }}
-        JIRA_PROD_URL: "https://sonarsource.atlassian.net/"
-        JIRA_SANDBOX_URL: "https://sonarsource-sandbox-608.atlassian.net/"
         INPUT_JIRA_PROJECT: ${{ inputs.jira-project }}
         INPUT_EPIC_NAME_PATTERN: ${{ inputs.epic-name-pattern }}
-        INPUT_JIRA_URL: ${{ ((inputs.use-jira-sandbox || env.USE_JIRA_SANDBOX) == 'true') && 'https://sonarsource-sandbox-608.atlassian.net/' || 'https://sonarsource.atlassian.net/' }}
+        INPUT_JIRA_URL: "https://sonarsource${{ ((inputs.use-jira-sandbox || env.USE_JIRA_SANDBOX) == 'true') && '-sandbox-608' || '' }}.atlassian.net/"
       run: |
         python ${{ github.action_path }}/resolve_ktlo_epic.py \
           --jira-project "$INPUT_JIRA_PROJECT" \
